requirements.txt
loguru
pip
python-dotenv
ruff
tqdm
typer
ipython
ipykernel
pandas
numpy
imbalanced-learn==0.12.3
scikit-learn==1.6.1
scikit-image
opencv-python
seaborn
imageio
scipy
six
pycocotools
ultralytics==8.3.228
weaviate-client
langchain
optuna
xgboost
sdv
shap
langchain-text-splitters
sentence-transformers
markdownify
beautifulsoup4
openai
cohere
fastapi
uvicorn
